+++ src/ParseNatSpec.js
